+++ src/main/java/frc/robot/BuildConstants.java
@@ -5,12 +5,12 @@ public final class BuildConstants {
     public static final String MAVEN_GROUP = "";
     public static final String MAVEN_NAME = "Flash-2023";
     public static final String VERSION = "unspecified";
-    public static final int GIT_REVISION = 168;
-    public static final String GIT_SHA = "b1d67f33b96e3bbe0dab3327d5bb97e70e6dc2af";
-    public static final String GIT_DATE = "2023-01-28 12:01:21 CST";
-    public static final String GIT_BRANCH = "main";
-    public static final String BUILD_DATE = "2023-01-28 12:11:01 CST";
-    public static final long BUILD_UNIX_TIME = 1674929461103L;
+    public static final int GIT_REVISION = 169;
+    public static final String GIT_SHA = "3e57fb81d5b6860d13c6d5663d5fe4a80854df83";
+    public static final String GIT_DATE = "2023-01-28 12:13:02 CST";
+    public static final String GIT_BRANCH = "Jonathan";
+    public static final String BUILD_DATE = "2023-01-28 15:54:23 CST";
+    public static final long BUILD_UNIX_TIME = 1674942863725L;
     public static final int DIRTY = 1;
 
     private BuildConstants() {}
